requirements.txt
torch==2.0.1
torchvision==0.15.2
hydra-core
omegaconf
transformers==4.49.0
diffusers==0.29.0
sentencepiece
opencv-python
peft==0.13.2
pyrootutils
xformers==0.0.22
accelerate==1.3.0
transformers_stream_generator
tqdm
notebook
numpy==1.21.2
huggingface_hub==0.29.3
timm>=0.9.16
attrdict
einops
